+++ src/main/java/dev/ftb/packcompanion/features/forcedgamerule/ForcedGameRulesFeature.java
@@ -1,0 +1,53 @@
+package dev.ftb.packcompanion.features.forcedgamerule;
+
+import dev.ftb.packcompanion.config.PCCommonConfig;
+import dev.ftb.packcompanion.core.Feature;
+import net.minecraft.nbt.ByteTag;
+import net.minecraft.nbt.IntTag;
+import net.minecraft.nbt.Tag;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.world.level.GameRules;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.server.ServerStartedEvent;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ForcedGameRulesFeature extends Feature.Common {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ForcedGameRulesFeature.class);
+
+    public ForcedGameRulesFeature(IEventBus modEventBus, ModContainer container) {
+        super(modEventBus, container);
+
+        NeoForge.EVENT_BUS.addListener(this::onLevelLoad);
+    }
+
+    public void onLevelLoad(ServerStartedEvent event) {
+        var forcedGameRules = PCCommonConfig.GAME_RULE_MAPPING.get();
+
+        MinecraftServer server = event.getServer();
+        GameRules.visitGameRuleTypes(new GameRules.GameRuleTypeVisitor() {
+            @Override
+            public <T extends GameRules.Value<T>> void visit(GameRules.Key<T> key, GameRules.Type<T> type) {
+                if (forcedGameRules.containsKey(key.getId())) {
+                    Tag value = forcedGameRules.get(key.getId());
+                    if (value instanceof ByteTag byteTag) {
+                        GameRules.BooleanValue rule = (GameRules.BooleanValue) server.getGameRules().getRule(key);
+                        var asBoolean = byteTag.getAsByte() != 0;
+                        if (rule.get() != asBoolean) {
+                            LOGGER.info("Setting (bool) game rule '{}' to {}", key.getId(), asBoolean);
+                            rule.set(asBoolean, server);
+                        }
+                    } else if (value instanceof IntTag intTag) {
+                        GameRules.IntegerValue rule = (GameRules.IntegerValue) server.getGameRules().getRule(key);
+                        if (rule.get() != intTag.getAsInt()) {
+                            LOGGER.info("Setting (int) game rule '{}' to {}", key.getId(), intTag.getAsInt());
+                            rule.set(intTag.getAsInt(), server);
+                        }
+                    }
+                }
+            }
+        });
+    }
+}